adding suspended and activated eventHandlers
